refactored to create HadoopJobUtils.proxyUserKillAllSpawnedHadoopJobs.
