+++ plugins/jobtype/src/azkaban/jobtype/HadoopPigJob.java
@@ -382,26 +382,6 @@ public class HadoopPigJob extends JavaProcessJob {
             getId());
     info("log file path is: " + logFilePath);
 
-    Properties properties = new Properties();
-    properties.putAll(jobProps.getFlattened());
-
-    try {
-      if (HadoopSecureWrapperUtils.shouldProxy(properties)) {
-        UserGroupInformation proxyUser =
-            HadoopSecureWrapperUtils.setupProxyUser(properties,
-                tokenFile.getAbsolutePath(), getLog());
-        proxyUser.doAs(new PrivilegedExceptionAction<Void>() {
-          @Override
-          public Void run() throws Exception {
-            HadoopJobUtils.killAllSpawnedHadoopJobs(logFilePath, getLog());
-            return null;
-          }
-        });
-      } else {
-        HadoopJobUtils.killAllSpawnedHadoopJobs(logFilePath, getLog());
-      }
-    } catch (Throwable t) {
-      warn("something happened while trying to kill all spawned jobs", t);
-    }
+    HadoopJobUtils.proxyUserKillAllSpawnedHadoopJobs(logFilePath, jobProps, tokenFile, getLog());
   }
 }
